Removed GoConfigMigration.upgradeIfNecessary(configFile, version) and its usages from tests. Moved the integration test cases to be written around GoConfigMigrator.migrate()
