+++ server/src/test-fast/java/com/thoughtworks/go/config/GoConfigMigratorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 ThoughtWorks, Inc.
+ * Copyright 2019 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,7 +51,7 @@ public class GoConfigMigratorTest {
         initMocks(this);
 
         goConfigMigrator = new GoConfigMigrator(goConfigMigration, systemEnvironment, fullConfigSaveNormalFlow, loader,
-                reader, configRepository, serverHealthService);
+                reader, configRepository, serverHealthService, null);
     }
 
     @Test
